output_files/crewai/screenplay_writer/pyproject.toml
[project]
name = "ai_crewforscreenwriting"
version = "0.1.0"
description = "Team that orchestrates the sequence of agents to convert a newsgroup post into a"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.10,<=3.13"

dependencies = [
    "crewai[tools]>=0.152.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
]

[project.scripts]
ai_crewforscreenwriting = "main:run"
train      = "main:train"

[tool.uv]
dev-dependencies = []

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"
